requirements/Tesrequirements.txt
amqp==1.4.9
aniso8601==1.2.0
anyjson==0.3.3
Babel==2.3.4
billiard==3.3.0.23
bleach==1.4.3
blinker==1.4
celery==3.1.23
click==6.6
dominate==2.2.1
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-Celery-Helper==1.1.0
Flask-HTTPAuth==3.2.1
Flask-JSON==0.3.0
flask-lesscss==0.9.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Moment==0.5.1
Flask-PageDown==0.2.2
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Table==0.2.12
Flask-WTF==0.12
ForgeryPy==0.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
Markdown==2.6.7
MarkupSafe==0.23
mysql-connector==2.1.4
pymongo==3.3.1
PyMySQL==0.7.9
python-dateutil==2.6.0
pytz==2016.6.1
six==1.10.0
SQLAlchemy==1.0.15
visitor==0.1.3
Werkzeug==0.11.11
WTForms==2.1
